Move folders based on documentation recommendation.
