+++ app/page.tsx
@@ -1,6 +1,6 @@
-import PageSection from "@/components/PageSection";
-import { getPage } from "@/utils/contentful";
-import { PageSectionType } from "@/constants/types";
+import PageSection from "@/app/components/PageSection";
+import { getPage } from "@/app/utils/contentful";
+import { PageSectionType } from "@/app/constants/types";
 
 // This is the root page to the website.
 export default async function Home() {
